+++ features/user/topchat/src/main/java/com/tokopedia/topchat/chatroom/view/presenter/TopChatRoomPresenter.kt
@@ -9,7 +9,7 @@ import com.tokopedia.atc_common.data.model.request.AddToCartOccRequestParams
 import com.tokopedia.atc_common.domain.model.response.AddToCartDataModel
 import com.tokopedia.atc_common.domain.usecase.AddToCartOccUseCase
 import com.tokopedia.atc_common.domain.usecase.AddToCartUseCase
-import com.tokopedia.attachproduct.resultmodel.ResultProduct
+import com.tokopedia.attachcommon.data.ResultProduct
 import com.tokopedia.chat_common.data.ChatroomViewModel
 import com.tokopedia.chat_common.data.ImageUploadViewModel
 import com.tokopedia.chat_common.data.ReplyChatViewModel
@@ -511,7 +511,7 @@ class TopChatRoomPresenter @Inject constructor(
     }
 
     override fun getShopFollowingStatus(
-            shopId: Int,
+            shopId: Long,
             onError: (Throwable) -> Unit,
             onSuccessGetShopFollowingStatus: (Boolean) -> Unit
     ) {
@@ -660,8 +660,8 @@ class TopChatRoomPresenter @Inject constructor(
         )
     }
 
-    override fun loadAttachmentData(msgId: Int, chatRoom: ChatroomViewModel) {
-        if (chatRoom.hasAttachment() && msgId != 0) {
+    override fun loadAttachmentData(msgId: Long, chatRoom: ChatroomViewModel) {
+        if (chatRoom.hasAttachment() && msgId != 0L) {
             chatAttachmentUseCase.getAttachments(
                     msgId, chatRoom.attachmentIds, ::onSuccessGetAttachments, ::onErrorGetAttachments
             )
